feat: reimplement core parts
